+++ app/src/components/Trends/Examples/TrendsExamplesItemChart.js
@@ -1,14 +1,9 @@
-import React, { Fragment } from 'react'
+import React from 'react'
 import { graphql } from 'react-apollo'
 import moment from 'moment'
-import { Line } from 'react-chartjs-2'
 import { trendsExploreGQL } from '../trendsExploreGQL'
-import {
-  mergeDataSourcesForChart,
-  parseTrendsGQLProps,
-  SourceColor
-} from '../trendsUtils'
-import PropTypes from 'prop-types'
+import { parseTrendsGQLProps } from '../trendsUtils'
+import TrendsChart from '../TrendsChart'
 
 const chartOptions = {
   animation: false,
@@ -35,74 +30,16 @@ const chartOptions = {
   }
 }
 
-const datasetOptions = {
-  borderColor: 'rgba(255, 193, 7, 1)',
-  borderWidth: 2,
-  pointRadius: 0,
-  fill: false
-}
-
-const propTypes = {
-  sources: PropTypes.object.isRequired,
-  topic: PropTypes.string.isRequired
-}
-
-const createDatasetOptions = (label, borderColor) => ({
-  label,
-  yAxisID: 'y-axis-0',
-  borderColor,
-  borderWidth: 0,
-  pointRadius: 0,
-  fill: false
-})
-
-const chartDatasetOptions = {
-  merged: createDatasetOptions('Merged', SourceColor['merged']),
-  telegram: createDatasetOptions('Telegram', SourceColor['telegram']),
-  reddit: createDatasetOptions('Reddit', SourceColor['reddit']),
-  professionalTradersChat: createDatasetOptions(
-    'Professional Traders Chat',
-    SourceColor['professionalTradersChat']
-  )
-}
-
-const composeSourcesChartDatasets = (sources, selectedSources) => {
-  if (selectedSources.includes('merged')) {
-    return [
-      {
-        data: [...mergeDataSourcesForChart(sources).values()],
-        ...chartDatasetOptions['merged']
-      }
-    ]
-  }
-  // AS ternary
-  return selectedSources.map(selectedSource => ({
-    data:
-      sources[selectedSource] &&
-      sources[selectedSource].map(item => item.mentionsCount),
-    ...chartDatasetOptions[selectedSource]
-  }))
-}
-
 const TrendsExamplesItemChart = ({ sources, selectedSources }) => {
-  const isLoading = !sources
-
-  const mergedSources = mergeDataSourcesForChart(sources)
-
-  const dataset = {
-    labels: [...mergedSources.keys()],
-    datasets: composeSourcesChartDatasets(sources, selectedSources)
-  }
-
   return (
-    <Fragment>
-      {isLoading && <div className='chart-loading-msg'>Loading...</div>}
-      <Line options={chartOptions} data={dataset} />
-    </Fragment>
+    <TrendsChart
+      sources={sources}
+      selectedSources={selectedSources}
+      chartOptions={chartOptions}
+      height={null}
+    />
   )
 }
-
-TrendsExamplesItemChart.propTypes = propTypes
 
 export default graphql(trendsExploreGQL, {
   props: parseTrendsGQLProps,
